move upload file classes into ui.upload namespace
