+++ nicegui/elements/upload_files.py
@@ -12,7 +12,7 @@ import anyio
 from starlette.datastructures import UploadFile
 from starlette.formparsers import MultiPartParser
 
-from . import json, run
+from .. import json, run
 
 
 @dataclass
@@ -119,8 +119,7 @@ def _cleanup_path(path: Path) -> None:
     path.unlink(missing_ok=True)
 
 
-async def create_file_upload(upload: UploadFile, *,
-                             chunk_size: int = 1024 * 1024) -> FileUpload:
+async def create_file_upload(upload: UploadFile, *, chunk_size: int = 1024 * 1024) -> FileUpload:
     """Create a file upload from a Starlette UploadFile.
 
     :param upload: the Starlette UploadFile to create a file upload from
